Updated the text processors to allow output to a different file from the input file.  And other tweaks.

+++ src/FileSharper/FileSharperCore/Processors/ProcessorBase.cs
@@ -5,12 +5,15 @@
 using System;
 using System.IO;
 using System.Threading;
+using FileSharperCore.Util;
 using Microsoft.VisualBasic.FileIO;
 
 namespace FileSharperCore.Processors
 {
     public abstract class ProcessorBase : PluggableItemBase, IProcessor
     {
+
+        public const string ORIGINAL_FILE_PATH = @"{DirectoryName}\{NameWithoutExtension}{Extension}";
 
         public InputFileSource InputFileSource { get; set; }
 
@@ -25,15 +28,48 @@ namespace FileSharperCore.Processors
 
         }
 
-        public void CopyAndDeleteTempFile(string tmpFile, string outFile, bool moveOriginalToRecycleBin)
+        public bool CopyAndDeleteTempFile(string tmpFile, string outFile, bool overwrite, bool moveOriginalToRecycleBin)
         {
-            if (moveOriginalToRecycleBin)
+            bool copied = false;
+            if (moveOriginalToRecycleBin && File.Exists(outFile))
             {
                 FileSystem.DeleteFile(outFile, UIOption.OnlyErrorDialogs, RecycleOption.SendToRecycleBin,
                 UICancelOption.DoNothing);
+                copied = true;
             }
-            File.Copy(tmpFile, outFile, true);
-            File.Delete(tmpFile);
+            try
+            {
+                File.Copy(tmpFile, outFile, overwrite);
+                copied = true;
+            }
+            catch (IOException)
+            {
+                
+            }
+            try
+            {
+                File.Delete(tmpFile);
+            }
+            catch (Exception)
+            {
+
+            }
+            return copied;
+        }
+
+        public ProcessingResult GetProcessingResultFromCopyAndDeleteTempFile(FileInfo file, string preReplaceOutFileName,
+            string tempFileName, bool overwrite, bool moveOriginalToRecycleBin)
+        {
+            string outfile = ReplaceUtil.Replace(preReplaceOutFileName, file);
+            bool wroteFile = CopyAndDeleteTempFile(tempFileName, outfile, overwrite, moveOriginalToRecycleBin);
+            if (wroteFile)
+            {
+                return new ProcessingResult(ProcessingResultType.Success, "Success", new FileInfo[] { new FileInfo(outfile) });
+            }
+            else
+            {
+                return new ProcessingResult(ProcessingResultType.Failure, "Failure", new FileInfo[] { });
+            }
         }
     }
 }
